модуляция js файлов
